Remove duplicate max_allocation_is_big_enough test utility function
