+++ zebra-chain/src/sapling/tests/preallocate.rs
@@ -14,11 +14,11 @@ use crate::{
         },
         PerSpendAnchor, SharedAnchor,
     },
-    serialization::{TrustedPreallocate, ZcashSerialize},
+    serialization::{tests::max_allocation_is_big_enough, TrustedPreallocate, ZcashSerialize},
 };
 
 use proptest::prelude::*;
-use std::{cmp::max, convert::TryInto};
+use std::cmp::max;
 
 proptest! {
     /// Confirm that each `Spend<PerSpendAnchor>` takes exactly
@@ -207,38 +207,4 @@ proptest! {
         // greatest upper bound across spends and outputs.
         prop_assert!((largest_allowed_vec_len as u64) <= max(SpendPrefixInTransactionV5::max_allocation(), OutputPrefixInTransactionV5::max_allocation()));
     }
-}
-
-/// Return the following calculations on `item`:
-///   smallest_disallowed_vec_len
-///   smallest_disallowed_serialized_len
-///   largest_allowed_vec_len
-///   largest_allowed_serialized_len
-fn max_allocation_is_big_enough<T>(item: T) -> (usize, usize, usize, usize)
-where
-    T: TrustedPreallocate + ZcashSerialize + Clone,
-{
-    let max_allocation: usize = T::max_allocation().try_into().unwrap();
-    let mut smallest_disallowed_vec = Vec::with_capacity(max_allocation + 1);
-    for _ in 0..(max_allocation + 1) {
-        smallest_disallowed_vec.push(item.clone());
-    }
-    let smallest_disallowed_serialized = smallest_disallowed_vec
-        .zcash_serialize_to_vec()
-        .expect("Serialization to vec must succeed");
-    let smallest_disallowed_vec_len = smallest_disallowed_vec.len();
-
-    // Create largest_allowed_vec by removing one element from smallest_disallowed_vec without copying (for efficiency)
-    smallest_disallowed_vec.pop();
-    let largest_allowed_vec = smallest_disallowed_vec;
-    let largest_allowed_serialized = largest_allowed_vec
-        .zcash_serialize_to_vec()
-        .expect("Serialization to vec must succeed");
-
-    (
-        smallest_disallowed_vec_len,
-        smallest_disallowed_serialized.len(),
-        largest_allowed_vec.len(),
-        largest_allowed_serialized.len(),
-    )
 }
