+++ zebra-chain/src/serialization/tests.rs
@@ -1,2 +1,4 @@
 mod preallocate;
 mod prop;
+
+pub use preallocate::max_allocation_is_big_enough;
